// src/api/model/examSubmission.ts
/**
 * Generated by orval v6.26.0 🍺
 * Do not edit manually.
 * student-hub-api
 * Api Documentation for StudentHub API
 * OpenAPI spec version: 0.0
 */
import type { ExamSubmissionGradeDistributionItem } from './examSubmissionGradeDistributionItem';

export interface ExamSubmission {
  createdAt?: string;
  criteriaPdf?: string;
  examId?: number;
  examName?: string;
  examPdf?: string;
  grade?: number;
  gradeDistribution?: ExamSubmissionGradeDistributionItem[];
  solutionPdf?: string;
  timeElaped?: number;
  userAnswers?: string[];
}